Basic MVC flow for adding listing and adding

+++ src/main/java/ie/philb/springtodo/service/TodoService.java
@@ -15,7 +15,7 @@ import java.util.List;
 
 public interface TodoService {
 
-    Todo getTodoById(long id);
+    Todo getTodoById(int id);
 
     Todo save(Todo todo);
